Replace CAS with funflow2 version
Replaced the top-level cas directory with the version from funflow2.
