+++ cas/store/changelog.md
@@ -1,0 +1,4 @@
+# Version 1.1.0
+
+- monad-control is no longer a dependency. cas-store is now using unliftio for
+  the same effect
