+++ cas/store/test/Test.hs
@@ -1,10 +1,12 @@
 import qualified CAS.ContentStore
-import           Test.Tasty
+import Test.Tasty
 
 main :: IO ()
 main = defaultMain tests
 
 tests :: TestTree
-tests = testGroup "Unit Tests"
-  [ CAS.ContentStore.tests
-  ]
+tests =
+  testGroup
+    "Unit Tests"
+    [ CAS.ContentStore.tests
+    ]
