--- funflow2/cas/hashable/cas-hashable.cabal
@@ -1,45 +1,0 @@
-Name:                cas-hashable
-Version:             1.0.1
-Synopsis:            A hashing class for content-addressed storage
-Description:
-            A hashing class for content-addressed storage. Content can be hashed either by value or through an IO action. Part of the funflow ecosystem.
-License:             MIT
-License-file:        LICENSE
-Author:              Tom Nielsen, Nicholas Clarke, Andreas Herrmann
-Maintainer:          yves.pares@tweag.io
-build-type:          Simple
-Cabal-Version:       >= 1.10
-homepage:            https://github.com/tweag/funflow
-bug-reports:         https://github.com/tweag/funflow
-category:            Control
-Tested-With:         GHC == 7.8.4, GHC == 7.10.2, GHC == 7.10.3, GHC == 8.0.1
-
-extra-source-files:
-                   changelog.md
-
-Library
-   ghc-options:       -Wall -fno-warn-type-defaults
-   hs-source-dirs:    src
-   default-language:  Haskell2010
-
-   Exposed-modules:  Data.CAS.ContentHashable
-   Build-depends:
-                 base                    >= 4.6 && <5
-               , aeson                   >= 1.2.3.0
-               , bytestring
-               , clock
-               , containers
-               , cryptonite
-               , ghc-prim
-               , hashable
-               , integer-gmp
-               , memory
-               , path                    > 0.6.0
-               , path-io
-               , safe-exceptions
-               , scientific
-               , text
-               , time
-               , unix
-               , unordered-containers
-               , vector
