--- funflow2/cas/s3/cas-hashable-s3.cabal
@@ -1,34 +1,0 @@
-Name:                cas-hashable-s3
-Version:             1.0.0
-Synopsis:            ContentHashable instances for S3 objects
-Description:
-            Provides ContentHashable instances for S3 objects
-License:             MIT
-License-file:        LICENSE
-Author:              Tom Nielsen, Nicholas Clarke, Andreas Herrmann
-Maintainer:          yves.pares@tweag.io
-build-type:          Simple
-Cabal-Version:       >= 1.10
-homepage:            https://github.com/tweag/funflow
-bug-reports:         https://github.com/tweag/funflow
-category:            Control
-Tested-With:         GHC == 7.8.4, GHC == 7.10.2, GHC == 7.10.3, GHC == 8.0.1
-
-extra-source-files:
-                   changelog.md
-
-Library
-   ghc-options:       -Wall -fno-warn-type-defaults
-   hs-source-dirs:    src
-   default-language:  Haskell2010
-
-   Exposed-modules:  Data.CAS.ContentHashable.S3
-   Build-depends:
-                 base                    >= 4.6 && <5
-               , aeson
-               , aws
-               , cas-hashable            >= 1.0.0 && < 2
-               , constraints
-               , http-conduit
-               , reflection
-               , resourcet
