--- funflow2/cas/store/cas-store.cabal
@@ -1,81 +1,0 @@
-Name:                cas-store
-Version:             1.1.0
-Synopsis:            A content-addressed storage
-Description:
-            A content-addressed storage supporting a remote caching. The API mainly consists of the cacheKleisliIO function which takes a (a -> m b) function
-            and runs it only if the store doesn't already contain a result for it. Part of the funflow ecosystem.
-License:             MIT
-License-file:        LICENSE
-Author:              Tom Nielsen, Nicholas Clarke, Andreas Herrmann
-Maintainer:          yves.pares@tweag.io
-build-type:          Simple
-Cabal-Version:       >= 1.10
-homepage:            https://github.com/tweag/funflow
-bug-reports:         https://github.com/tweag/funflow
-category:            Control
-Tested-With:         GHC == 7.8.4, GHC == 7.10.2, GHC == 7.10.3, GHC == 8.0.1
-
-extra-source-files:
-                   changelog.md
-
-Library
-   ghc-options:       -Wall -fno-warn-type-defaults
-   hs-source-dirs:    src
-   default-language:  Haskell2010
-
-   Exposed-modules:  Data.CAS.ContentStore
-                   , Data.CAS.Lock
-                   , Data.CAS.RemoteCache
-                   , Data.CAS.StoreOrphans
-   Other-modules: Data.CAS.ContentStore.Notify
-   Build-depends:
-                 base                    >= 4.6 && <5
-               , aeson                   >= 1.2.3.0
-               , async
-               , bytestring
-               , cas-hashable            >= 1.0.1 && < 2
-               , containers
-               , cryptonite
-               , directory
-               , filepath
-               , hashable
-               , hostname
-               , lens
-               , path
-               , path-io
-               , random
-               , safe-exceptions
-               , sqlite-simple
-               , store
-               , tar
-               , text
-               , unix
-               , unliftio
-   if os(linux)
-     CPP-options: -DOS_Linux
-     Other-modules: Data.CAS.ContentStore.Notify.Linux
-     Build-depends: hinotify             >= 0.3.9
-   else
-     if os(darwin) || os(freebsd)
-       CPP-options: -DOS_BSD
-       Other-modules: Data.CAS.ContentStore.Notify.BSD
-       Build-depends: kqueue
-
-Test-suite unit-tests
-  type:               exitcode-stdio-1.0
-  default-language:   Haskell2010
-  hs-source-dirs:     test
-  main-is:            Test.hs
-  other-modules:      CAS.ContentStore
-  ghc-options:        -Wall -threaded
-  build-depends:      base
-                    , async
-                    , cas-hashable
-                    , cas-store
-                    , containers
-                    , path
-                    , path-io
-                    , safe-exceptions
-                    , tasty
-                    , tasty-hunit
-                    , unix
