--- funflow2/cas/store/changelog.md
@@ -1,4 +1,0 @@
-# Version 1.1.0
-
-- monad-control is no longer a dependency. cas-store is now using unliftio for
-  the same effect
