--- funflow2/cas/store/test/Test.hs
@@ -1,12 +1,0 @@
-import qualified CAS.ContentStore
-import Test.Tasty
-
-main :: IO ()
-main = defaultMain tests
-
-tests :: TestTree
-tests =
-  testGroup
-    "Unit Tests"
-    [ CAS.ContentStore.tests
-    ]
